src/main/resources/scripts/python/jmcomic/requirements.txt
jmcomic~=2.6.4
pika~=1.3.2
img2pdf
plugin_jm_server
pillow~=11.3.0